file structure update, break down large files
